requirements.txt
Flask==2.1.2
psycopg2-binary==2.9.3
Flask-SQLAlchemy===2.5.1
Flask-Migrate==3.1.0
PyJWT==2.4.0
Flask-API